backend/requirements.txt
# Minimal backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pymongo==4.6.0
slowapi==0.1.9
loguru==0.7.2

# Optional LDAP support (install if needed)
# ldap3==2.9.1